Renamed folders
Close #55
