--- src/ecstasy/integration/CMakeLists.txt
@@ -1,8 +1,0 @@
-set(INCROOT ${INCROOT}/integration)
-set(SRCROOT ${SRCROOT}/integration)
-
-add_subdirectory(event)
-
-set(SRC 
-    ${SRC}
-    PARENT_SCOPE)
