+++ src/ecstasy/integrations/CMakeLists.txt
@@ -1,0 +1,8 @@
+set(INCROOT ${INCROOT}/integrations)
+set(SRCROOT ${SRCROOT}/integrations)
+
+add_subdirectory(event)
+
+set(SRC 
+    ${SRC}
+    PARENT_SCOPE)
