+++ src/ecstasy/integrations/event/EventsManager.hpp
