+++ src/ecstasy/registry/Registry.hpp
@@ -21,10 +21,10 @@
 #include "ecstasy/query/Query.hpp"
 #include "ecstasy/query/Select.hpp"
 #include "ecstasy/query/modifiers/Modifier.hpp"
-#include "ecstasy/resource/entity/Entities.hpp"
-#include "ecstasy/storage/IStorage.hpp"
-#include "ecstasy/storage/Instances.hpp"
-#include "ecstasy/storage/StorageConcepts.hpp"
+#include "ecstasy/resources/entity/Entities.hpp"
+#include "ecstasy/storages/IStorage.hpp"
+#include "ecstasy/storages/Instances.hpp"
+#include "ecstasy/storages/StorageConcepts.hpp"
 #include "ecstasy/system/ISystem.hpp"
 #include "util/Allocator.hpp"
 #include "util/meta/outer_join.hpp"
