+++ src/ecstasy/resources/Resource.hpp
