+++ src/ecstasy/resources/entity/RegistryEntity.hpp
