+++ src/ecstasy/storages/IStorage.hpp
