+++ tests/entity/tests_RegistryEntity.cpp
@@ -1,7 +1,7 @@
 #include <gtest/gtest.h>
-#include "ecstasy/resource/entity/Entities.hpp"
-#include "ecstasy/resource/entity/RegistryEntity.hpp"
-#include "ecstasy/storage/MapStorage.hpp"
+#include "ecstasy/resources/entity/Entities.hpp"
+#include "ecstasy/resources/entity/RegistryEntity.hpp"
+#include "ecstasy/storages/MapStorage.hpp"
 
 struct Position {
     int x;
